+++ common/src/main/java/earth/terrarium/cadmus/common/claims/CadmusDataHandler.java
@@ -6,6 +6,9 @@ import earth.terrarium.cadmus.api.teams.TeamProviderApi;
 import it.unimi.dsi.fastutil.ints.IntIntPair;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.StringTag;
+import net.minecraft.nbt.Tag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.level.block.Block;
@@ -56,10 +59,10 @@ public class CadmusDataHandler extends SaveHandler {
         if (tag.contains("allowedBlocks")) {
             CompoundTag allowedBlocksTag = tag.getCompound("allowedBlocks");
             allowedBlocksTag.getAllKeys().forEach(id -> {
-                CompoundTag blockTag = allowedBlocksTag.getCompound(id);
+                ListTag blockTag = allowedBlocksTag.getList(id, Tag.TAG_STRING);
                 Set<ResourceLocation> blocks = new HashSet<>();
-                blockTag.getAllKeys().forEach(namespace ->
-                    blocks.add(new ResourceLocation(namespace, blockTag.getString(namespace))));
+                blockTag.forEach(tagEntry ->
+                    blocks.add(new ResourceLocation(tagEntry.getAsString())));
                 allowedBlocks.put(id, blocks);
             });
         }
@@ -98,8 +101,8 @@ public class CadmusDataHandler extends SaveHandler {
 
         CompoundTag allowedBlocksTag = new CompoundTag();
         this.allowedBlocks.forEach((id, blocks) -> {
-            CompoundTag blockTag = new CompoundTag();
-            blocks.forEach(block -> blockTag.putString(block.getNamespace(), block.getPath()));
+            ListTag blockTag = new ListTag();
+            blocks.forEach(block -> blockTag.add(StringTag.valueOf(block.toString())));
             allowedBlocksTag.put(id, blockTag);
         });
         tag.put("allowedBlocks", allowedBlocksTag);